requirements.txt
python-telegram-bot==13.13
redis==3.2.1
python-dotenv==1.0.0
requests==2.28.2